Able to update the database when we edit a job application.
The page isn't refreshing after we edit an application cause it to show the unedited application, so that needs to be fixed.

Also, going to need to handle the update of the cv ad cl. But before that, i need to change the design so we can add a new cv/cl.